requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
numpy==1.26.4
pillow==10.4.0
tensorflow==2.16.2
matplotlib==3.9.1
tensorflow_datasets==4.9.6
opencv-python-headless==4.10.0.84
